+++ .gitignore
@@ -73,7 +73,9 @@ data
 bs_pics
 bs_gifs
 bs_data
-bg_model_pics
 bg_model_gifs
+bg_model_pics
+bg_subtract_pics
+bg_subtract_gifs
 kalman_filter_gifs
 kalman_filter_pics
