+++ MLX90640/config.py
@@ -4,5 +4,5 @@ godec_pics_path = "godec_pics/"
 godec_gifs_path = "godec_gifs/"
 bs_pics_path = "bs_pics/"
 bs_results_path = "bs_results/"
-bg_model_pics_path = "bg_model_pics/"
-bg_model_gifs_path = "bg_model_gifs/"
+bg_subtraction_pics_path = "bg_subtract_pics/"
+bg_subtraction_gifs_path = "bg_subtract_gifs/"
